#71 💷 refactor: useExchange 커스텀 hooks 생성
