+++ src/hooks/useExchange.ts
@@ -1,0 +1,34 @@
+import { useEffect } from "react";
+import { useSearchParams } from "next/navigation";
+
+import { useQuery } from "@tanstack/react-query";
+import axios from "axios";
+import { useRecoilValue, useSetRecoilState } from "recoil";
+
+import { getUserData } from "@/hooks/getUserData";
+import { userId, userUidAssetData } from "@/atoms/atom";
+
+const useExchange = () => {
+  const setUserAssetData = useSetRecoilState(userUidAssetData);
+  const userUid = useRecoilValue(userId);
+
+  const searchParams = useSearchParams();
+
+  const market_code = searchParams.get("market_code");
+
+  const { data } = useQuery({
+    queryKey: ["currentPrice"],
+    queryFn: () =>
+      axios(`https://api.upbit.com/v1/ticker?markets=${market_code}`),
+  });
+
+  const currentPrice = data?.data[0].trade_price;
+
+  useEffect(() => {
+    getUserData(userUid, setUserAssetData);
+  }, []);
+
+  return { currentPrice };
+};
+
+export default useExchange;
